Split api files out by name instead of by provider
