+++ src/utils/__tests__/toasts.spec.js
@@ -8,7 +8,7 @@ jest.mock('react-redux-toastr', () => ({
   },
 }));
 
-jest.mock('../../store/api/firebase', () => ({
+jest.mock('../../store/api/entity', () => ({
   uploadSector: jest.fn(() => Promise.resolve()),
 }));
